remove unneeded cast

+++ cmd/rb-scrobbler.go
@@ -90,7 +90,7 @@ Automatically ("keep", "delete" or "delete-on-success") at end of program`)
 	}
 
 	/* When given a file, start executing here */
-	offsetSec := float64(*offset * SECONDS_IN_HOUR)
+	offsetSec := *offset * SECONDS_IN_HOUR
 
 	tracks, err := logFile.ImportLog(logPath, offsetSec, colours)
 	if err != nil {
